디자이너 상세 조회 모달 작업중

+++ Customer/src/main/java/com/h2/dogrooming/product/Product.java
@@ -14,7 +14,7 @@ public class Product {
 
     @Id
     @GeneratedValue
-    private Integer productId;
+    private long productId;
 
     @Column(length = 40, nullable = false)
     private String adminId;
